app/requirements.txt
requests==2.28.2
flask==2.2.2
Flask-Cors==4.0.0
# In order to be compatible with itsdangerous==1.1.0 needed by flask-oidc 1.4.0 which is latest version
# Flask==1.1.2
flask-smorest==0.40.0
# pytest==7.2.1
######## ENV vars ########
python-dotenv==0.21.1
######## ORM ########
sqlalchemy==2.0.1
flask-sqlalchemy==3.0.3
######## For Database ########
# flask-migrate
psycopg2-binary==2.9.6
######## For server ########
# gunicorn
######## Keycloak ########
# flask-oidc==1.4.0
# itsdangerous==1.1.0
# flask-oidc-ext==1.4.5
# itsdangerous==2.1.2
python-keycloak==3.1.1
######## For Authentication ########
# flask-jwt-extended==4.4.4
passlib==1.7.1